screen size based lod switch

+++ Snooper/Rendering/Components/Primitive/PrimitiveDescriptor.cs
@@ -6,7 +6,7 @@ public unsafe struct PrimitiveDescriptor(CullingBounds bounds)
     
     public fixed uint LOD_FirstIndex[Settings.MaxNumberOfLods];
     public fixed uint LOD_BaseVertex[Settings.MaxNumberOfLods];
-    public fixed uint LOD_ScreenSize[Settings.MaxNumberOfLods];
+    public fixed float LOD_ScreenSize[Settings.MaxNumberOfLods];
     public fixed uint LOD_SectionCount[Settings.MaxNumberOfLods];
     public fixed uint LOD_SectionOffset[Settings.MaxNumberOfLods];
 }
